requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
django-rest-auth-bearer==0.1.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
lxml==4.6.2
psycopg2==2.8.6
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
zeep==4.0.0
